requirements.txt
lark-parser>=0.12.0
Jinja2>=2.11
pytest>=6.0
amitools>=0.5.0
Pillow>=8.0